requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
stripe==7.0.0
jinja2==3.1.2
aiofiles==23.2.1
email-validator==2.1.0
pydantic[email]==2.5.0

